Split the kvm ut test cases to correspond to different platforms
    Split the kvm ut test cases to pass unit-tests on Arm64.
    I will add the tls and full-context test cases for Arm64 later.
