+++ pkg/sentry/platform/kvm/BUILD
@@ -60,6 +60,7 @@ go_library(
 go_test(
     name = "kvm_test",
     srcs = [
+        "kvm_amd64_test.go",
         "kvm_test.go",
         "virtual_map_test.go",
     ],
